updated dem udfs
