+++ community/sina/DEM_Tile_Example/DEM_Tile_Example.py
@@ -1,17 +1,17 @@
 def udf(
     bounds: fused.types.Bounds = [-77.083,38.804,-76.969,38.983],
-    provider: str = "AWS"
+    provider: str = "MSPC"
 ):
     import odc.stac
     import palettable
     import planetary_computer
     import pystac_client
     from pystac.extensions.eo import EOExtension as eo
-    import utils
 
+    common = fused.load("https://github.com/fusedio/udfs/tree/b7637ee/public/common/")
     # Load pinned versions of utility functions.
-    common_utils = fused.load("https://github.com/fusedio/udfs/tree/2f41ae1/public/common/").utils
-    tile = common_utils.get_tiles(bounds, clip=True)
+
+    tile = common.get_tiles(bounds, clip=True)
     zoom = tile.iloc[0].z
 
     print(f'{type(tile) = }')
@@ -56,7 +56,7 @@ def udf(
     arr = xarray_dataset["data"].max(dim="time")
 
     # Visualize that data as an RGB image.
-    rgb_image = utils.visualize(
+    rgb_image = common.visualize(
         data=arr,
         min=0,
         max=3000,
