+++ community/sina/DEM_Tile_Example/meta.json
@@ -57,41 +57,22 @@
             "fused:udfType": "raster",
             "fused:slug": "DEM_Tile_Example",
             "fused:name": "DEM_Tile_Example",
-            "fused:id": null,
-            "fused:gitUrl": "https://github.com/fusedio/udfs/tree/e3202f5807f0fb399abc4965d3ae35cc1c48715d/public/DEM_Tile_Example/",
-            "fused:gitShortUrl": "https://github.com/fusedio/udfs/tree/e3202f5/public/DEM_Tile_Example/",
-            "fused:gitPath": "public/DEM_Tile_Example",
-            "fused:gitRef": "e3202f5807f0fb399abc4965d3ae35cc1c48715d",
+            "fused:gitPath": "community/sina/DEM_Tile_Example",
+            "fused:gitRef": "2528576155696ad0aa24beef28643891343f2211",
             "fused:gitAuthorNames": [
-              "Sina Kashuk",
-              "Tyler Erickson",
-              null,
-              "Plinio Guzman",
-              "Isaac Brodsky"
+              null
             ],
             "fused:gitAuthorUsernames": [
-              "kashuk",
-              "tylere",
-              "charlesfused",
-              "pgzmnk",
-              "isaacbrodsky"
+              "MaxLenormand"
             ],
             "fused:gitAuthorUrls": [
-              "https://github.com/kashuk",
-              "https://github.com/tylere",
-              "https://github.com/charlesfused",
-              "https://github.com/pgzmnk",
-              "https://github.com/isaacbrodsky"
+              "https://github.com/MaxLenormand"
             ],
             "fused:gitAuthorAvatarUrls": [
-              "https://avatars.githubusercontent.com/u/12432685?v=4",
-              "https://avatars.githubusercontent.com/u/92862?v=4",
-              "https://avatars.githubusercontent.com/u/148997493?v=4",
-              "https://avatars.githubusercontent.com/u/27398253?v=4",
-              "https://avatars.githubusercontent.com/u/9139378?v=4"
+              "https://avatars.githubusercontent.com/u/44133329?v=4"
             ],
-            "fused:gitLastModified": "2024-10-11T21:33:13+00:00",
-            "fused:explorerTab": "public",
+            "fused:gitLastModified": "2025-07-23T14:53:34+00:00",
+            "fused:explorerTab": "self",
             "fused:gitRepo": "fusedio/udfs",
             "fused:assetUrl": "https://fused-magic.s3.us-west-2.amazonaws.com/thumbnails/udfs-staging/DEM_Tile_Example.png",
             "fused:tags": [
@@ -113,15 +94,12 @@
               }
             ],
             "fused:description": "#### Overview\nThe Copernicus DEM is a digital surface model (DSM), which represents the surface of the Earth including buildings, infrastructure, and vegetation. This DSM is based on radar satellite data acquired during the TanDEM-X Mission, which was funded by a public-private partnership between the German Aerospace Centre (DLR) and Airbus Defence and Space.\n\n\nCopernicus DEM is available at both 30-meter and 90-meter resolution.\n\n\nSee the [Product Handbook](https://object.cloud.sdsc.edu/v1/AUTH_opentopography/www/metadata/Copernicus_metadata.pdf) for more information.\n\nSee the dataset page on [OpenTopography](https://doi.org/10.5069/G9028PQB).\n#### STAC Collections\n*  [AWS Earth Search](https://earth-search.aws.element84.com/v1/collections/cop-dem-glo-30)\n* [Microsoft Planetary Computer](https://planetarycomputer.microsoft.com/api/stac/v1/collections/cop-dem-glo-30)\n\n\n## Run this in any Jupyter Notebook\n\n```python\nimport fused\nimport geopandas as gpd\n\nudf_url = \"https://github.com/fusedio/udfs/tree/main/public/DEM_Tile_Example\"\nudf = fused.load(udf_url)\n\nbounds = gpd.GeoDataFrame.from_features(\n    {\"type\": \"FeatureCollection\", \"features\": [{\"id\": \"0\", \"type\": \"Feature\", \"properties\": {\"x\": 1311, \"y\": 3164, \"z\": 13}, \"geometry\": {\"type\": \"Polygon\", \"coordinates\": [[[-124.1455078125, 41.77131167976406], [-124.1455078125, 41.73852846935915], [-124.189453125, 41.73852846935915], [-124.189453125, 41.77131167976406], [-124.1455078125, 41.77131167976406]]]}}]}\n)\n\narr = udf.run_local(bounds=bounds)\narr\n```\n",
-            "fused:explorerPin": "13"
+            "fused:explorerPin": "13",
+            "fused:userCreatedShareToken": "fsh_6qDvivboWoWGuPx4iaaw8y",
+            "fused:id": null
           },
           "source": "DEM_Tile_Example.py",
-          "headers": [
-            {
-              "module_name": "utils",
-              "source_file": "utils.py"
-            }
-          ]
+          "headers": []
         }
       }
     ],
