--- community/sina/DEM_Tile_Example/utils.py
@@ -1,4 +1,0 @@
-# Load utility functions.
-visualize = fused.load(
-    "https://github.com/fusedio/udfs/tree/5cfb808/public/common/"
-).utils.visualize
